Changed tooling and added routes
- Set python version to 3.13 and uv to 0.9.5 via .tool-versions
- Migrated package to uv
- Used openspec to generate a feature spec
- Used Droid to implement the features which were basically to add
endpoints that our frontend will connect to that cover the basics:
    - upload images
    - send images to AI provider
    - receive images from AI provider
    - send images to B2
- Create tests
- Some errors that were relatively minor after Droid was done:
    - Dependency issue with psycopg2, just need to uv add
    - Mock URLs needed to either be updated to real errors or
temporarily disabled.
- Droid was using GPT5 Codex, but in order to debug, GLM4.6 has been
pretty good lately so it went to work and got main.py up and running
again.
- There was some error with the postgresql connection string to Railway.
- There were database type/importing errors but GLM4.6 took care of it.
Not sure if it dropped all the data from the database, but there's not
too much anyways so it shouldn't be a big deal to re-import if
necessary.
- Had GLM4.6 document everything it worked on and fixed.
- Had GLM4.6 document the basic structure of the flow of the backend
which a frontend agent should be able to use.